move analytics out of the exports
